+++ CHANGELOG.md
@@ -13,3 +13,15 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - TODO: implement support for make file usage
 - TODO: implement run behavior
 - TODO: add build directory for output
+
+## [0.0.2] - 2019-12-30
+
+- minor settings tweaks
+
+## [0.0.3] - 2019-12-31
+
+- more settings tweaks - renaming/reorganizing
+- add quick config for common emulator settings
+- added quick config support for Altirra
+- added launch emulator for just starting the configured emulator
+